fix(checker): dedupe repeated object-literal keys in the source-type display
`object_literal_source_type_display` builds the source side of a TS2322 /
TS2345 from the object-literal syntax rather than from the finalized literal
type, and emitted one member per element. A repeated key therefore rendered
once per occurrence — `{ zprop: number; wprop: number; zprop: number; }` where
`tsc` prints `{ zprop: number; wprop: number; }`.

Structural rule: when an object literal repeats a direct key, tsc's property
table keeps the FIRST declaration's position and takes the LAST declaration's
value; tsz now does the same in the diagnostic display builder, which is the
layer that owns this rendering.

The finalized object-literal type was always correct here — it is keyed by
`Atom` in a map — which is why the same literal deduped correctly when routed
through a variable or an argument position and diverged only on the annotated
initializer path that uses this AST-order reconstruction. Same shape as
#16011: two subsystems computing one fact, the less-informed one winning.

Found by resurrecting #16013's orphaned-test backlog (below); the two failing
assertions were verified against tsc 7.0.2 before either side was touched.

test(ci): register #16013's 6 remaining orphaned test files and empty the baseline

All six had `#[test]` fns that were never compiled into any target under
`autotests = false`, so they never ran. Registered via `[[test]]` stanzas:

  tsz-checker/tests/indexed_access_concrete_missing_key_ts2536_tests.rs  11/11
  tsz-checker/tests/ts6133_write_only_destructuring_tests.rs              8/8
  tsz-checker/tests/generic_mapped_to_index_signature_tests.rs            6/6
  tsz-checker/tests/object_literal_synthetic_this_display_order_tests.rs  5/5
  tsz-emitter/tests/async_method_super_capture_3759_tests.rs              6/6
  tsz-cli/tests/skip_lib_check_declaration_diagnostics_tests.rs           2/2

`ts6133_write_only_destructuring_tests.rs` did not even compile — it read a
`Diagnostic::line_and_char` field that does not exist. Because the file was in
no target, `cargo check` never saw it either. Rewritten onto the canonical
`test_utils::diagnostic_line_column` helper.

`scripts/ci/orphaned-test-files-baseline.txt` goes 6 entries -> 0. The file
stays so the shrink-only guard keeps something to diff against.
