+++ scripts/ci/orphaned-test-files-baseline.txt
@@ -11,14 +11,25 @@
 # cross-crate `crates/tsz-core/src/lib.rs` `#[path]` include into
 # `tsz-checker/tests/` — the same mechanism #16013 itself credited for
 # excluding `constructor_accessibility.rs`/`void_return_exception.rs`, just
-# not run for these five. The remaining 6 files (38 #[test] fns) are
-# genuinely unreachable; each wants oracle-first verification before being
-# resurrected (a dormant suite freezes stale expectations — see #16013's
-# #15632-revert precedent), so they are tracked here rather than fixed in
-# this diff.
-tsz-checker/tests/indexed_access_concrete_missing_key_ts2536_tests.rs
-tsz-checker/tests/ts6133_write_only_destructuring_tests.rs
-tsz-checker/tests/generic_mapped_to_index_signature_tests.rs
-tsz-checker/tests/object_literal_synthetic_this_display_order_tests.rs
-tsz-emitter/tests/async_method_super_capture_3759_tests.rs
-tsz-cli/tests/skip_lib_check_declaration_diagnostics_tests.rs
+# not run for these five.
+#
+# The remaining 6 files (38 #[test] fns) are now registered via `[[test]]`
+# stanzas, run, and green:
+#
+#   tsz-checker/tests/indexed_access_concrete_missing_key_ts2536_tests.rs  11/11
+#   tsz-checker/tests/ts6133_write_only_destructuring_tests.rs              8/8
+#   tsz-checker/tests/generic_mapped_to_index_signature_tests.rs            6/6
+#   tsz-checker/tests/object_literal_synthetic_this_display_order_tests.rs  5/5
+#   tsz-emitter/tests/async_method_super_capture_3759_tests.rs              6/6
+#   tsz-cli/tests/skip_lib_check_declaration_diagnostics_tests.rs           2/2
+#
+# Resurrection was oracle-first, per #16013's own instruction and the
+# a0bca422f2 -> 4521dc46e7 revert precedent. Two of the 38 did fail, and the
+# oracle said the *tests* were right: duplicate object-literal keys were
+# rendering once per occurrence in the annotated-initializer source display.
+# That was a real product bug and was fixed at the display builder rather than
+# rewritten away.
+#
+# The baseline is now empty. It stays in the tree so the guard keeps a
+# shrink-only file to diff against, and so a newly orphaned file still fails
+# the per-merge lane instead of silently joining a list.
